+++ api/api.go
@@ -409,9 +409,3 @@ func FmtValue(value interface{}, depth int) string {
 
 	return fmtReflectValue(reflect.ValueOf(value), depth)
 }
-
-// CaptchaForce api method.
-func (vk *VK) CaptchaForce(params Params) (response int, err error) {
-	err = vk.RequestUnmarshal("captcha.force", params, &response)
-	return
-}
