+++ api/captcha.go
@@ -1,0 +1,7 @@
+package api
+
+// CaptchaForce api method.
+func (vk *VK) CaptchaForce(params Params) (response int, err error) {
+	err = vk.RequestUnmarshal("captcha.force", params, &response)
+	return
+}
